Edit my_led Fixed my_delay Add basic my_menu

+++ Project/Driver/My_GPIO.c
@@ -1,5 +1,5 @@
 /**
- * Passed
+ * Test: Passed
  */
 #include "stm32f4xx_gpio.h"
 #include "my_gpio.h"
@@ -7,8 +7,8 @@
 /**
  * Init GPIO
  */
-void my_gpio_init (int GPIOx, GPIOMode_TypeDef mode, int pinx, GPIOOType_TypeDef type
-                , GPIOSpeed_TypeDef speed, GPIOPuPd_TypeDef pupd) {
+void my_gpio_init (uint16_t GPIOx, GPIOMode_TypeDef mode, uint16_t pinx, GPIOOType_TypeDef type,
+                   GPIOSpeed_TypeDef speed, GPIOPuPd_TypeDef pupd) {
 
     GPIO_InitTypeDef gpio_init;
     gpio_init.GPIO_Pin = pinx;
@@ -23,45 +23,37 @@ void my_gpio_init (int GPIOx, GPIOMode_TypeDef mode, int pinx, GPIOOType_TypeDef
             RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOA, ENABLE);
             GPIO_Init(GPIOA, &gpio_init);
             break;
-        
         case MY_GPIOB:
             RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOB, ENABLE);
-            GPIO_Init(GPIOB, &gpio_init);             
-            break;                                    
-                                                      
-        case MY_GPIOC:                                
+            GPIO_Init(GPIOB, &gpio_init);
+            break;
+        case MY_GPIOC:
             RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOC, ENABLE);
-            GPIO_Init(GPIOC, &gpio_init);             
-            break;                                    
-                                                      
-        case MY_GPIOD:                                
+            GPIO_Init(GPIOC, &gpio_init);
+            break;
+        case MY_GPIOD:
             RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOD, ENABLE);
-            GPIO_Init(GPIOD, &gpio_init);             
-            break;                                    
-                                                      
-        case MY_GPIOE:                                
+            GPIO_Init(GPIOD, &gpio_init);
+            break;
+        case MY_GPIOE:
             RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOE, ENABLE);
-            GPIO_Init(GPIOE, &gpio_init);             
-            break;                                    
-                                                      
-        case MY_GPIOF:                                
+            GPIO_Init(GPIOE, &gpio_init);
+            break;
+        case MY_GPIOF:
             RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOF, ENABLE);
-            GPIO_Init(GPIOF, &gpio_init);             
-            break;                                    
-                                                      
-        case MY_GPIOG:                                
+            GPIO_Init(GPIOF, &gpio_init);
+            break;
+        case MY_GPIOG:
             RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOG, ENABLE);
-            GPIO_Init(GPIOG, &gpio_init);             
-            break;                                    
-                                                      
-        case MY_GPIOH:                                
+            GPIO_Init(GPIOG, &gpio_init);
+            break;
+        case MY_GPIOH:
             RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOH, ENABLE);
-            GPIO_Init(GPIOH, &gpio_init);             
-            break;                                    
-                                                      
-        case MY_GPIOI:                                
+            GPIO_Init(GPIOH, &gpio_init);
+            break;
+        case MY_GPIOI:
             RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOI, ENABLE);
-            GPIO_Init(GPIOI, &gpio_init);             
-            break;                                    
+            GPIO_Init(GPIOI, &gpio_init);
+            break;
     }
 }
